requirements.txt
basedosdados==2.0.3
google-cloud-bigquery-storage>=2.24
pandas>=2.0
openpyxl
python-dotenv
plotly>=5.0
nbformat>=4.2.0
